make buildExecutionContext a static method on Executor
makes execute file even thinner
