+++ src/execution/executor.ts
@@ -5,7 +5,9 @@ import type { Maybe } from '../jsutils/Maybe';
 import { inspect } from '../jsutils/inspect';
 import { memoize2 } from '../jsutils/memoize2';
 import { invariant } from '../jsutils/invariant';
+import { devAssert } from '../jsutils/devAssert';
 import { isPromise } from '../jsutils/isPromise';
+import { isObjectLike } from '../jsutils/isObjectLike';
 import { promiseReduce } from '../jsutils/promiseReduce';
 import { promiseForObject } from '../jsutils/promiseForObject';
 import { addPath, pathToArray } from '../jsutils/Path';
@@ -15,10 +17,12 @@ import { GraphQLError } from '../error/GraphQLError';
 import { locatedError } from '../error/locatedError';
 
 import type {
+  DocumentNode,
   OperationDefinitionNode,
   FieldNode,
   FragmentDefinitionNode,
 } from '../language/ast';
+import { Kind } from '../language/kinds';
 
 import type { GraphQLSchema } from '../type/schema';
 import type {
@@ -32,6 +36,7 @@ import type {
   GraphQLTypeResolver,
   GraphQLList,
 } from '../type/definition';
+import { assertValidSchema } from '../type/validate';
 import {
   SchemaMetaFieldDef,
   TypeMetaFieldDef,
@@ -49,10 +54,18 @@ import { getOperationRootType } from '../utilities/getOperationRootType';
 
 import { isAsyncIterable } from '../jsutils/isAsyncIterable';
 
-import type { ExecutionContext, ExecutionResult } from './execute';
-import { getArgumentValues } from './values';
+import type {
+  ExecutionArgs,
+  ExecutionContext,
+  ExecutionResult,
+} from './execute';
+import { getVariableValues, getArgumentValues } from './values';
 import { collectFields } from './collectFields';
 import { mapAsyncIterator } from './mapAsyncIterator';
+
+export interface BuildExecutionContextArgs extends ExecutionArgs {
+  subscribeFieldResolver?: Maybe<GraphQLFieldResolver<any, any>>;
+}
 
 /**
  * This class is exported only to assist people in implementing their own executors
@@ -109,6 +122,93 @@ export class Executor {
     this._typeResolver = typeResolver;
     this._subscribeFieldResolver = subscribeFieldResolver;
     this._errors = errors;
+  }
+
+  /**
+   * Constructs a ExecutionContext object from the arguments passed to
+   * execute, which we will pass throughout the other execution methods.
+   *
+   * Throws a GraphQLError if a valid execution context cannot be created.
+   *
+   * @internal
+   */
+  static buildExecutionContext(
+    args: BuildExecutionContextArgs,
+  ): ReadonlyArray<GraphQLError> | ExecutionContext {
+    const {
+      schema,
+      document,
+      rootValue,
+      contextValue,
+      variableValues: rawVariableValues,
+      operationName,
+      fieldResolver,
+      typeResolver,
+      subscribeFieldResolver,
+    } = args;
+
+    // If arguments are missing or incorrect, throw an error.
+    assertValidExecutionArguments(schema, document, rawVariableValues);
+
+    let operation: OperationDefinitionNode | undefined;
+    const fragments: ObjMap<FragmentDefinitionNode> = Object.create(null);
+    for (const definition of document.definitions) {
+      switch (definition.kind) {
+        case Kind.OPERATION_DEFINITION:
+          if (operationName == null) {
+            if (operation !== undefined) {
+              return [
+                new GraphQLError(
+                  'Must provide operation name if query contains multiple operations.',
+                ),
+              ];
+            }
+            operation = definition;
+          } else if (definition.name?.value === operationName) {
+            operation = definition;
+          }
+          break;
+        case Kind.FRAGMENT_DEFINITION:
+          fragments[definition.name.value] = definition;
+          break;
+      }
+    }
+
+    if (!operation) {
+      if (operationName != null) {
+        return [
+          new GraphQLError(`Unknown operation named "${operationName}".`),
+        ];
+      }
+      return [new GraphQLError('Must provide an operation.')];
+    }
+
+    // istanbul ignore next (See: 'https://github.com/graphql/graphql-js/issues/2203')
+    const variableDefinitions = operation.variableDefinitions ?? [];
+
+    const coercedVariableValues = getVariableValues(
+      schema,
+      variableDefinitions,
+      rawVariableValues ?? {},
+      { maxErrors: 50 },
+    );
+
+    if (coercedVariableValues.errors) {
+      return coercedVariableValues.errors;
+    }
+
+    return {
+      schema,
+      fragments,
+      rootValue,
+      contextValue,
+      operation,
+      variableValues: coercedVariableValues.coerced,
+      fieldResolver: fieldResolver ?? defaultFieldResolver,
+      typeResolver: typeResolver ?? defaultTypeResolver,
+      subscribeFieldResolver,
+      errors: [],
+    };
   }
 
   /**
@@ -895,3 +995,90 @@ export function getFieldDef(
   }
   return parentType.getFields()[fieldName];
 }
+
+/**
+ * Essential assertions before executing to provide developer feedback for
+ * improper use of the GraphQL library.
+ *
+ * @internal
+ */
+export function assertValidExecutionArguments(
+  schema: GraphQLSchema,
+  document: DocumentNode,
+  rawVariableValues: Maybe<{ readonly [variable: string]: unknown }>,
+): void {
+  devAssert(document, 'Must provide document.');
+
+  // If the schema used for execution is invalid, throw an error.
+  assertValidSchema(schema);
+
+  // Variables, if provided, must be an object.
+  devAssert(
+    rawVariableValues == null || isObjectLike(rawVariableValues),
+    'Variables must be provided as an Object where each property is a variable value. Perhaps look to see if an unparsed JSON string was provided.',
+  );
+}
+
+/**
+ * If a resolveType function is not given, then a default resolve behavior is
+ * used which attempts two strategies:
+ *
+ * First, See if the provided value has a `__typename` field defined, if so, use
+ * that value as name of the resolved type.
+ *
+ * Otherwise, test each possible type for the abstract type by calling
+ * isTypeOf for the object being coerced, returning the first type that matches.
+ */
+export const defaultTypeResolver: GraphQLTypeResolver<unknown, unknown> =
+  function (value, contextValue, info, abstractType) {
+    // First, look for `__typename`.
+    if (isObjectLike(value) && typeof value.__typename === 'string') {
+      return value.__typename;
+    }
+
+    // Otherwise, test each possible type.
+    const possibleTypes = info.schema.getPossibleTypes(abstractType);
+    const promisedIsTypeOfResults = [];
+
+    for (let i = 0; i < possibleTypes.length; i++) {
+      const type = possibleTypes[i];
+
+      if (type.isTypeOf) {
+        const isTypeOfResult = type.isTypeOf(value, contextValue, info);
+
+        if (isPromise(isTypeOfResult)) {
+          promisedIsTypeOfResults[i] = isTypeOfResult;
+        } else if (isTypeOfResult) {
+          return type.name;
+        }
+      }
+    }
+
+    if (promisedIsTypeOfResults.length) {
+      return Promise.all(promisedIsTypeOfResults).then((isTypeOfResults) => {
+        for (let i = 0; i < isTypeOfResults.length; i++) {
+          if (isTypeOfResults[i]) {
+            return possibleTypes[i].name;
+          }
+        }
+      });
+    }
+  };
+
+/**
+ * If a resolve function is not given, then a default resolve behavior is used
+ * which takes the property of the source object of the same name as the field
+ * and returns it as the result, or if it's a function, returns the result
+ * of calling that function while passing along args and context value.
+ */
+export const defaultFieldResolver: GraphQLFieldResolver<unknown, unknown> =
+  function (source: any, args, contextValue, info) {
+    // ensure source is a value for which property access is acceptable.
+    if (isObjectLike(source) || typeof source === 'function') {
+      const property = source[info.fieldName];
+      if (typeof property === 'function') {
+        return source[info.fieldName](args, contextValue, info);
+      }
+      return property;
+    }
+  };
